% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{declareGlobs}
\alias{declareGlobs}
\title{declareGlobs()}
\usage{
declareGlobs(bigsnp = NULL, torus = NULL)
}
\arguments{
\item{bigsnp}{string denoting path to reference panel in rds format (converted from PLINK using bigsnpr::add_bed)}

\item{torus}{location of torus executable}
}
\description{
Sets global variants: location of torus executable and location of reference panel rds file
}
