% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_pips.R
\name{.calculate_pips}
\alias{.calculate_pips}
\title{FINEMAP-miss PIPs}
\usage{
.calculate_pips(
  init_configs,
  p,
  stored_init_config_sizes,
  stored_log_bf,
  stored_unique_configs,
  lse
)
}
\arguments{
\item{init_configs}{String vector of initial configurations for each iteration of
Shotgun Stochastic Search. Stored as a collapsed string.}

\item{p}{Number of variants.}

\item{stored_init_config_sizes}{Numeric vector of sizes of initial configurations for each iteration of
Shotgun Stochastic Search.}

\item{stored_log_bf}{Numeric vector of log Bayes-factors of evaluated configurations.}

\item{stored_unique_configs}{Binary Vector of denoting uniqueness of evaluated configuration.}

\item{lse}{logsumexp of evaluated configuration log Bayes-factors}
}
\value{
Numeric vector of PIPs for each variant.
}
\description{
Computes the postrior inclusion probabilities (PIPs), given the evaluated
causal configurations in FINEMAP-miss.
}
