% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_credible_sets.R
\name{.create_credible_sets}
\alias{.create_credible_sets}
\title{FINEMAP-MISS credible sets}
\usage{
.create_credible_sets(
  cred_sizes,
  post_k,
  max_causals,
  stored_log_bf,
  stored_config_sizes,
  stored_configs,
  log_prior,
  z_RMi_R,
  I_tR_RMi_R,
  comp_R,
  p
)
}
\arguments{
\item{cred_sizes}{Vector or numeric of size of credible sets to be created.
If NULL, then the rounded expectation of the posterior number of causal variants
is used.}

\item{post_k}{Vector of posterior probability of number of casual variants,
from 0:max_causals}

\item{max_causals}{Numeric, maximum number of causal variants}

\item{stored_log_bf}{Vector of log Bayes-factors computed during fine-mapping.}

\item{stored_config_sizes}{Vector of size of configurations
evaluated during fine-mapping.}

\item{stored_configs}{Vector of evaluated configurations, stored as a collapsed string.}

\item{log_prior}{Natural log transformed prior for number of causal variants.}

\item{z_RMi_R}{Pre-multiplied vector for fine-mapping.}

\item{I_tR_RMi_R}{Pre-multiplied matrix for fine-mapping.}

\item{comp_R}{LD matrix, with diagonal set to 0.}

\item{p}{Number of variants.}
}
\value{
List of credible sets as matrices for each 'cred_size' in 'cred_sizes'. If 'length(cred_sizes) == 1', then the
single credible set matrix is returned.
}
\description{
Computes credible sets from the fine-mapping output of FINEMAP-MISS.
}
