% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_credible_sets_FM.R
\name{.create_credible_sets_FM}
\alias{.create_credible_sets_FM}
\title{FINEMAP credible sets}
\usage{
.create_credible_sets_FM(
  cred_sizes,
  post_k,
  max_causals,
  stored_log_bf,
  stored_config_sizes,
  stored_configs,
  log_prior,
  comp_R,
  RM,
  R,
  p,
  Si_Ii,
  I2,
  z,
  tau
)
}
\arguments{
\item{cred_sizes}{Vector or numeric of size of credible sets to be created.
If NULL, then the rounded expectation of the posterior number of causal variants
is used.}

\item{post_k}{Vector of posterior probability of number of casual variants,
from 0:max_causals}

\item{max_causals}{Numeric, maximum number of causal variants}

\item{stored_log_bf}{Vector of log Bayes-factors computed during fine-mapping.}

\item{stored_config_sizes}{Vector of size of configurations
evaluated during fine-mapping.}

\item{stored_configs}{Vector of evaluated configurations, stored as a collapsed string.}

\item{log_prior}{Natural log transformed prior for number of causal variants.}

\item{comp_R}{LD matrix, with diagonal set to 0.}

\item{RM}{z-score covariance matrix}

\item{R}{LD matrix.}

\item{p}{Number of variants.}

\item{Si_Ii}{Diagonal matrix of standard error and square root of INFO score inverses.}

\item{I2}{Diagonal matrix of INFO scores.}

\item{z}{vector of z-scores.}

\item{tau}{prior standard deviation.}
}
\value{
List of credible sets as matrices for each 'cred_size' in 'cred_sizes'. If 'length(cred_sizes) == 1', then the
single credible set matrix is returned.
}
\description{
Computes credible sets from the fine-mapping output of the FINEMAP model.
}
