% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_RM_matrix.R
\name{.create_RM_matrix}
\alias{.create_RM_matrix}
\title{Create z-score covariance matrix}
\usage{
.create_RM_matrix(
  ses,
  betas = NULL,
  INFO,
  R,
  beta_shrink = T,
  n_studies,
  p,
  estimate_M = FALSE,
  max_overlap = T
)
}
\arguments{
\item{ses}{A vector or matrix of GWAS marginal effect standard errors.
Takes the form of a vector if there is only a single dataset or the data has
been combined in advance. Otherwise, it is a matrix where each column contains
the standard errors for one dataset. For any unobserved variants in a study,
the standard errors should be set to Inf.}

\item{betas}{A vector or matrix of GWAS marginal effects.
Takes the form of a vector if there is only a single dataset or the data has
been combined in advance. Otherwise, it is a matrix where each column contains
the marginal effects for one dataset. For any unobserved variants in a study,
the marginal effects should be set to 0.}

\item{INFO}{A vector or matrix of variant imputation scores.
Takes the form of a vector if there is only a single dataset or the data has
been combined in advance. Otherwise, it is a matrix where each column contains
the marginal effects for one dataset. For any unobserved variants in a study,
the marginal effects should be set to 0.}

\item{R}{Reference LD matrix for the set of analyzed variants.}

\item{beta_shrink}{Binary parameter denoting whether beta shrinkage in the
marginal effect covariances is applied.}

\item{n_studies}{Number of studies. Should be equal to the number of columns in
ses abnd betas. If the data has been combined in advance, set to 1.}

\item{p}{Number of variants.}

\item{estimate_M}{Binary parameter denoting whether the sample overlap matrix
\eqn{\boldsymbol{M}} should be estimated rather than directly computed. Should
be used when the data has been combined in advance.}

\item{max_overlap}{Binary parameter denoting whether the maximum sample
overlap is assumed when estimating \eqn{\boldsymbol{M}}.}
}
\value{
z-score covariance matrix \eqn{\boldsymbol{R_M}}
}
\description{
Computes the covariance matrix \eqn{\boldsymbol{R_M}} for the z-scores from a GWAS meta-analysis
with possible missing information.
}
