% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_logbf.R
\name{.eval_logbf}
\alias{.eval_logbf}
\title{Evaluate log Bayes-factor}
\usage{
.eval_logbf(z_RMi_R, I_tR_RMi_R, configuration)
}
\arguments{
\item{z_RMi_R}{Pre-multiplied vector}

\item{I_tR_RMi_R}{Pre-multiplied matrix}

\item{configuration}{Index vector for the current configuration}
}
\value{
Numeric value of the log Bayes factor for the configuration
}
\description{
Evaluate the log Bayes-factor of a configuration in FINEMAP-MISS. Computed using
the pre-computed vector \eqn{\boldsymbol{\widehat z}^{T} \boldsymbol{R_M}^{-1}\boldsymbol{R}_{\boldsymbol{\gamma}} }
and matrix \eqn{\mathbb{I}_p + \boldsymbol{R}_{\boldsymbol{\gamma}}^{T} \boldsymbol{R_M}^{-1} \boldsymbol{R}_{\boldsymbol{\gamma}} }
}
