% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_dmvnorm.R
\name{.log_dmvnorm}
\alias{.log_dmvnorm}
\title{Log-dmvnorm
evaluate the log-density of a multivariate normal distribution.}
\usage{
.log_dmvnorm(
  x,
  mu = rep(0, length(x)),
  S = diag(1, length(x)),
  pivot = F,
  cholesky = T
)
}
\arguments{
\item{x}{vector, point to be evaluated.}

\item{mu}{vector, distribution mean}

\item{S}{matrix, distribution covariance}

\item{pivot}{should pivoting be used for cholseky decomposition?}

\item{cholesky}{is cholesky decomposition used for inversion? If not base R solve is used.}
}
\value{
log-density of the observed point x.
}
\description{
Log-dmvnorm
evaluate the log-density of a multivariate normal distribution.
}
