% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_new_configuration.R
\name{.sample_new_config}
\alias{.sample_new_config}
\title{Sample new configuration, Stochastic Shotgun Search (SSS)}
\usage{
.sample_new_config(init_config, init_configs, prob, n_causals, p, non_causals)
}
\arguments{
\item{init_config}{Numeric vector of the current initial configuration.}

\item{init_configs}{String vector of all initial configurations, collapsed into strings.}

\item{prob}{Probability vector for sampling the next initial configuration.}

\item{n_causals}{Number of causal configurations in init_config, 'length(init_config)'}

\item{p}{Number of variants.}

\item{non_causals}{Set of variants not in 'init_config'.}
}
\value{
Initial configuration for the next iteration of SSS.
}
\description{
Given the fully evaluated neighborhood of configurations in a iteration in SSS,
this function samples the configuration that will be used as the initial
configuration in the next SSS iteration.
}
