% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torus.R
\name{findTorus}
\alias{findTorus}
\title{findTorus}
\usage{
findTorus(path = NULL)
}
\arguments{
\item{path}{String denoting the full path to torus if known}
}
\value{
String; torus path
}
\description{
Finds torus if it has been added to the PATH variable
}
