% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IVW.R
\name{IVW}
\alias{IVW}
\title{Inverse-variance weighting}
\usage{
IVW(betas, ses)
}
\arguments{
\item{betas}{Matrix of measurements. Each column contains the measurements
from one study. Any missing values from a dataset should be set to 0.}

\item{ses}{Matrix of measurement standard errors. Each column contains the standard errors
from one study. Any missing values from a dataset should be set to Inf.}
}
\value{
List containing the meta-analyzed measurements and standard errors.
}
\description{
Performs a fixed-effects meta-analysis given a set of measurements and standard errors
across multiple datasets.
}
\examples{
#example meta-analysis of two datasets. Data point 3 is missing in dataset 1.
betas <- cbind(c(0.5, 0.1, 0), c(0.4, 0.15, 0.25))
ses <- cbind(c(0.11, 0.1, Inf), c(0.12, 0.11, 0.1))
meta_analysis <- IVW(betas, ses)

}
