% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_FINEMAPMISS.R
\name{run_FINEMAPMISS}
\alias{run_FINEMAPMISS}
\title{Fine-mapping meta-analyses with missing information}
\usage{
run_FINEMAPMISS(
  betas,
  ses,
  R,
  n_studies,
  variant_sample_sizes = NULL,
  freqs,
  INFO = NULL,
  tau = 0.05,
  adj = 1e-04,
  n_reps = 50,
  prob_threshold = 0.001,
  max_causals = 5,
  cred_sizes = NULL,
  quiet = F,
  adj_match = T,
  RM = NULL,
  estimate_M = FALSE,
  max_overlap = T,
  scaled_data = F,
  use_N = T,
  INFO_prior = T,
  rsid = NULL,
  position = NULL,
  allele1 = NULL,
  allele2 = NULL,
  chromosome = NULL,
  export_configs = FALSE,
  init_config = NULL,
  find_optimal_start = FALSE,
  n_start_configs = 1000,
  start_config_sizes = 100,
  initial_config_size = 2
)
}
\arguments{
\item{betas}{A vector or matrix of GWAS marginal effects \eqn{\boldsymbol{\widehat \beta}}.
Takes the form of a vector if there is only a single dataset or the data has
been combined in advance. Otherwise, it is a matrix where each column contains
the marginal effects for one dataset. For any unobserved variants in a study,
the marginal effects should be set to \code{0}.}

\item{ses}{A vector or matrix of GWAS marginal effect standard errors \eqn{\boldsymbol{s}}.
Takes the form of a vector if there is only a single dataset or the data has
been combined in advance. Otherwise, it is a matrix where each column contains
the standard errors for one dataset. For any unobserved variants in a study,
the standard errors should be set to \code{Inf}.}

\item{R}{Reference LD matrix \eqn{\boldsymbol{R}} for the set of analyzed variants.}

\item{n_studies}{Number of studies. Should be equal to the number of columns in
ses abnd betas. If the data has been combined in advance, set to \code{1}.}

\item{variant_sample_sizes}{Vector of combined sample sizes per variant, from the meta-analysis.}

\item{freqs}{A vector or matrix of variant frequencies.
Takes the form of a vector if there is only a single dataset or the data has
been combined in advance. Otherwise, it is a matrix where each column contains
the variant frequencies for one dataset. For any unobserved variants in a study,
the marginal effects should be set to \code{0}.}

\item{INFO}{A vector or matrix of variant INFO scores.
Takes the form of a vector if there is only a single dataset or the data has
been combined in advance. Otherwise, it is a matrix where each column contains
the INFO scores for one dataset. For any unobserved variants in a study,
the marginal effects should be set to \code{0}.}

\item{tau}{Prior standard error for the casual effects. Set by default to \code{0.05}.}

\item{adj}{Adjustment constant \eqn{\varepsilon}, shrinks the correlations in
\eqn{\boldsymbol{R_M}} to make its inversion stable.}

\item{n_reps}{Maximum number of Stochastic Shotgun Search (SSS) iterations.}

\item{prob_threshold}{SSS termination threshold, based on probability mass added at each iteration.}

\item{max_causals}{Maximum number of causal variants.}

\item{cred_sizes}{Vector or numeric of size of credible sets to be evaluated. If no value
is provided, the rounded expected posterior for number of causal variants is used.}

\item{quiet}{Should the current progress of SSS be repressed?}

\item{adj_match}{Should the \eqn{\boldsymbol{R}} be adjusted in a similar way to \eqn{\boldsymbol{R_M}} when computing log Bayes-factors?}

\item{RM}{If \eqn{\boldsymbol{R_M}} has been computed in advance, it can be provided as input.}

\item{estimate_M}{Should \eqn{\boldsymbol{M}} be estimated? This should be used if data has been meta-analyzed in advance.}

\item{max_overlap}{Binary parameter denoting whether the maximum sample
overlap is assumed when estimating \eqn{\boldsymbol{M}}.}

\item{scaled_data}{Has the data been scaled with allele frequencies in advance?}

\item{use_N}{Should the variant sample sizes be used for fine-mapping, instead of the
GWAS standard errors? (experimental)}

\item{INFO_prior}{Should the INFO scores be used to scale the prior? (experimental)}

\item{rsid}{rsid vector (optional identifier).}

\item{position}{position vector (optional identifier).}

\item{allele1}{reference allele vector (optional identifier).}

\item{allele2}{alternate allele vector (optional identifier).}

\item{chromosome}{chromosome vector (optional identifier).}

\item{init_config}{Initial configuration, from which fine-mapping is started.}

\item{find_optimal_start}{Should a search be made to find an optimal starting configuration for fine-mapping?}

\item{n_start_configs}{How many configurations should be checked to determine optimal start?}

\item{start_config_sizes}{How large are the initial start configurations?}

\item{initial_config_size}{Size of initial config}
}
\value{
A list of objects
\describe{
  \item{\code{summary_table}}{Data frame with entries:
  \itemize{
  \item{\code{rank}: The order in which the variants appear when sorted by PIP.}
  \item{\code{rsid}: Rsid (if supplied, NA otherwise).}
  \item{\code{chromosome}: Chromosome (if supplied, NA otherwise).}
  \item{\code{allele1}: Allele1 (if supplied, NA otherwise).}
  \item{\code{allele2}: Allele2 (if supplied, NA otherwise).}
  \item{\code{maf}: Minor allele frequency.}
  \item{\code{beta}: Marginal effect.}
  \item{\code{se}: Standard error.}
  \item{\code{z}: Z-score.}
  \item{\code{prob}: Posterior inclusion probability (PIP).}
  }}
  \item{\code{cred}}{List or matrix of credible sets.}
  \item{\code{post_k}}{Posterior probability distribution for the number of causal variants from \code{0:max_causals}}
  \item{\code{evaluated_configs}}{Optional output if \code{export_configs == TRUE}. Data frame of evaluated configurations and associated information.
  \itemize{
   \item{\code{configuration}: Causal configuration.}
   \item{\code{log_bf}: Causal configuration log Bayes-factor.}
   \item{\code{config_size}: Causal configuration size (how many causal variants?).}
   \item{\code{unique_config}: Is this a unique configuration?.}
  }}
}
}
\description{
\code{run_FINEMAPMISS} performs Bayesian variable selection on a set of genetic
variants and a phenotype, using marginal effect estimates \eqn{\boldsymbol{\widehat \beta}},
 their standard errors \eqn{\boldsymbol{s}}, and  a reference LD panel \eqn{\boldsymbol{R}}
 as input. \cr
 \code{FINEMAPMISS} differs from previous fine-mapping methods, as it allows for the combination
 of multiple datasets across biobanks, while also functioning with missing data. \cr
 If given separate datasets, this function will automatically perform the meta-analysis
 and fine-map the data. \cr
 If the data are combined into a single dataset in advance, \code{FINEMAPMISS}
 can still be run, but with reduced accuracy.
}
\examples{
#Loading toy data with one true causal variant.
data("toydata_FINEMAPMISS")

betas <- toydata_FINEMAPMISS$betas
ses <- toydata_FINEMAPMISS$ses
MAF <- toydata_FINEMAPMISS$MAF
LD <- toydata_FINEMAPMISS$LD

n <- toydata_FINEMAPMISS$study_sample_sizes
p <- dim(LD)[1]

#Simulating missingness in 20\% of variants (including the true causal variant)
missing_data <- unique(sort(c(toydata_FINEMAPMISS$causal_snp,
                              cbind(sample(1:p, round(p*0.2))))))

#Which dataset are the variants missing from?
which_dataset_missing <- sample(1:2, length(missing_data), replace = TRUE)

#Setting missing data to 0 or Inf.
betas[missing_data[which_dataset_missing == 1],1] <- 0
betas[missing_data[which_dataset_missing == 2],2] <- 0
ses[missing_data[which_dataset_missing == 1],1] <- Inf
ses[missing_data[which_dataset_missing == 2],2] <- Inf
MAF[missing_data[which_dataset_missing == 1],1] <- 0
MAF[missing_data[which_dataset_missing == 2],2] <- 0

#Setting variants sample sizes
variant_sample_sizes <- rep(sum(n), p)
variant_sample_sizes[missing_data] <- n[1]

#Running FINEMAPMISS
output_FMM <- FINEMAPMISS(ses = ses,
                         betas = betas,
                         R = LD,
                         n_studies = 2,
                         variant_sample_sizes = variant_sample_sizes,
                         freqs = MAF)


}
